refactor(backmp11): state machine signature
